requirements.txt
pymupdf==1.23.5
python-docx==1.1.0
spacy==3.7.2
networkx==3.2.1

